Documentation update for the uptime_checks lookup
Add details of data structure returned by the lookup to the RETURN
string used by Ansible.
Update the README with details of how to call ansible-doc to read the
documentation for this lookup. This means we don't have to document the
lookup twice, in the code itself and again in the README.
Also removed some debug and verbose output from the lookup which should
not have been left in the code.